requirements.txt
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
google-cloud-bigquery==3.13.0
google-auth==2.23.4
streamlit==1.28.1
plotly==5.17.0
openpyxl==3.1.2
kaggle==1.5.16
python-dotenv==1.0.0
schedule==1.2.0
logging==0.4.9.6
fastapi==0.104.1
uvicorn==0.24.0
flask==3.0.0
flask-cors==4.0.0
datetime
json
os
sys
time
random
math 